useEffect getting triggers moved to Triggercontext
